testVectorRotate works only after casting the result of 1x3matrixmultiplication to integers.

+++  saving-face --username 1900Geek@gmail.com/include/sf_tr_func.h
@@ -88,9 +88,9 @@ namespace SF
 
 	inline void MatrixMultiply1b3(SF_MODEL_COORD3D &model_r3_coord,const SF_R3_COORD &tr_Coord3d, const SF_TR_MATRIX &tr_matrix)
 	{
-		model_r3_coord.x = (tr_Coord3d.x * tr_matrix.rotMTX[0] + tr_Coord3d.y * tr_matrix.rotMTX[1] + tr_Coord3d.z * tr_matrix.rotMTX[2]);
-		model_r3_coord.y = (tr_Coord3d.x * tr_matrix.rotMTX[3] + tr_Coord3d.y * tr_matrix.rotMTX[4] + tr_Coord3d.z * tr_matrix.rotMTX[5]);
-		model_r3_coord.z = (tr_Coord3d.x * tr_matrix.rotMTX[6] + tr_Coord3d.y * tr_matrix.rotMTX[7] + tr_Coord3d.z * tr_matrix.rotMTX[8]);
+		model_r3_coord.x = (int)(tr_Coord3d.x * tr_matrix.rotMTX[0] + tr_Coord3d.y * tr_matrix.rotMTX[1] + tr_Coord3d.z * tr_matrix.rotMTX[2]);
+		model_r3_coord.y = (int)(tr_Coord3d.x * tr_matrix.rotMTX[3] + tr_Coord3d.y * tr_matrix.rotMTX[4] + tr_Coord3d.z * tr_matrix.rotMTX[5]);
+		model_r3_coord.z = (int)(tr_Coord3d.x * tr_matrix.rotMTX[6] + tr_Coord3d.y * tr_matrix.rotMTX[7] + tr_Coord3d.z * tr_matrix.rotMTX[8]);
 	}
 
 	inline void MatrixMultiply3b3(SF_SCALAR* out, const SF_SCALAR* in1,const SF_SCALAR* in2)
